+++ jumi-core/src/test/java/fi/jumi/core/runs/AssigningRunIdsTest.java
@@ -9,10 +9,8 @@ import fi.jumi.actors.ActorRef;
 import fi.jumi.api.drivers.*;
 import fi.jumi.core.output.OutputCapturer;
 import fi.jumi.core.runners.TestClassListener;
-import org.apache.commons.io.output.NullOutputStream;
 import org.junit.Test;
 
-import java.nio.charset.Charset;
 import java.util.*;
 import java.util.concurrent.*;
 
@@ -25,7 +23,7 @@ public class AssigningRunIdsTest {
 
     private final TestClassListener listener = mock(TestClassListener.class);
     private final RunIdSequence runIdSequence = new RunIdSequence();
-    private final OutputCapturer outputCapturer = new OutputCapturer(new NullOutputStream(), new NullOutputStream(), Charset.defaultCharset());
+    private final OutputCapturer outputCapturer = new OutputCapturer();
     private final SuiteNotifier notifier = new DefaultSuiteNotifier(ActorRef.wrap(listener), runIdSequence, outputCapturer);
 
     @Test