+++ RELEASE-NOTES.md
@@ -6,6 +6,7 @@ Release Notes
 
 - Removed from jumi-api's Javadocs mentions of planned features which do not yet exist
 - Jumi Actors is from this release onwards its own project with its own release cycle
+- Added a default constructor to `OutputCapturer` in order to avoid the need of a `NullOutputStream` in tests
 
 ### Jumi 0.1.196 (2012-09-19)